Removes global variable references from calculatedOneBP.pyx
I think the code generally will be improved by passing function arguments
around explicitly, instead of implicitly using global variables.. Initially this may
mean having a lot of arguments to some functions but as we continue to refactor we'll
hopefully be able to improve the situation.

The reason I think this will be helpful is because it makes "control flow" explicit
and easier to follow. It becomes clear what must be called in what order. It also
makes it easier to see what data each function needs, which make the functions
easier to understand and reason about.