Adds -rngSeed option to CorrectBias_stored
This option will eventually be added to CorrectBias as well.

To get repeatable runs I had to factor out the part of performRegression that called
np.random.choice because it's called from two different processes and the order would
be non-deterministic.

Additionally, I changed some variable names to be more descriptive, added a "label"
argument to performRegression so the output files of the process that finishes first
won't be clobbered by the output files of the process that finishes second.

Finally, I created some Cython classes to use instead of lists of 3-tuples for Training
Sets and Training Regions.