Refactors performRegression in CorrectBias_stored
I think this factoring makes the code easier to follow and understand. It's broken down into more
functions which allows for code reuse and the computing code is more clearly deliniated from the
graph plitting code.

In my testing the output is exactly the same as the code without these changes.