Improves speed of generateNormalizedObBWs ~10x
Running this function on my test dataset initially took about 25 minutes.
I changed the algorithm itself and used some more of the features of
numpy and pyBigWig. These changes brought the time down to ~15 minutes.

Then I moved the function into cython and made a lot of the scalar variables cdefs
and created a MemoryView for the two numpy arrays. These changes brought the time
down to ~2.5 minutes!

I don't have tests, but I compared the output, using diff, to a known good output
from an earlier run on the same data and there is no difference in the files.