Improves mean calculation
using the pyBigWig stats method to calculate means is pretty slow. Additionally,
the "binning" algorithm is... opaque.

This creates new methods for calculating the means and doing the binning which are currently
used for training set calculations. This makes mean calculation significantly faster than
using pyBigWig with more control over the binning algorithm.

Hopefully this machinery can also be used to replace binning and mean calculations in
calculating task covariates, which would provide significant speedup and simplify the
code quite a bit.

Note, however, that this breaks "backwards compatibility" with versions prior to this one (0.23). This means the same data input will result in a slightly different output.

Additionally, I made writes to the training set file (in "fillTrainingSetMeta") buffered.
Instead of writing every line directly to the file it writes them to a string buffer (which is much faster) and only writes the buffer to the file once.