[BUGFIX] Sections not filtered, coalescing wrong
Added a check to writeBedFile and coalesceSections when setting the
end of the last region to analysisEnd. The purpose of that check was
to not include data that we didn't want to analyize. But the check also
needed to account for the situation where analysisEnd was bigger than
than the end of the last region. That's what changing `x = analysEnd`
to `x = min(x, analysisEnd)` does.

I also noticed that coalesceSections, which I adapted from
generateNormalizedObBWs, still uses '1' instead of 'binSize' in a couple
of places. This corrects the oversight.