[BUGFIX] Creating multiple lists like [[]] * n is incorrect
This creates n copies of one list, not n lists. This means that
In [1]: a = [[]] * 5
In [2]: a[0].append(1)
In [3]: a
Out[3]: [[1], [1], [1], [1], [1]]

Which is not what we want.
In [1]: a = [[] for _ in range(5)]
In [2]: a[0].append(1)
In [3]: a
Out[3]: [[1], [], [], [], []]

is the desired behavior.

Pulls out the "rotate matrix of strings" code so it can be tested.