Renames FA/faFile/fa to genome
"fa" was short for "fasta", a file format. However, the data format CRADLE
expects is 2bit, so the name was misleading on that account.

Beyond the inaccuracy of the names, though, I don't think they described the
right thing. The file contains the sequences of the human genome and that's
what's important -- not the file format. So I've changed all the references
from fa/FA/etc. to genome/GENOME/etc.

The old "-faFile" command line argument still works, for backwards compatability,
but it's no longer documented and should be considered deprecated.