Adds test for alignCoordinatesToCovariateFileBoundaries
FKA alignCoordinatesToHDF

The test itself isn't much, beyond the creation of a new class mocking
a py2bit object.

However, it took a while to understand what all the pieces of the function were
and what all the magic numbers did. This led to changes throughout CorrectBiasStored
that, I think, improve the clarity of the code.

Also, I fixed some typos in the README and bin/cradle files.