Fixes a bigwig data order issue
This bigwig issue was that the order of the chromosomes in the header created
for the new bigwig files did not match the order that data was added to the
bigwig files. The chromosome data used to create the header was an unordered set,
so the order was more-or-less random.

The fix is to use a list instead of a set. The list gets updated when the set is added to or sorted.
That's not exactly true -- the list update is deferred until it's the list is needed.