Adds explicit "slots" to classes
According to https://pythonspeed.com/articles/python-object-memory/ and
https://docs.python.org/3.9/reference/datamodel.html#object.__slots__
Setting __slots__ will both decrease memory usage and potentially speed up
member access. In my testing a ChromoRegion only take about half the space it
did without __slots__. As there can be a lot of ChromoRegion, this is tens, if
not hundreds of megabytes of memory savings.

There are probably similar savings per object for ChromoRegionSet and StoredCovariates
but there aren't many of those objects, so it's not as big a win.