Pre-opens bigwig files in correctReadCounts.py
BigWig files get repeatedly opened and closed, literally thousands
or millions of times in particular in selectOverallIdx -- every bw
gets opened and closed on each region. This takes up time,
unnecessarily.

This change raises file-opening (and closing) to the outer-most possible
scope of `correctReadCount` so the opening and closing happens as
infrequently as possible.