Outputs covariate data to per-chrom hdf files
Splits the data up by chromosome and runs "calculateTaskCovariates" on each
chromosome of data. This allows the output to match the input CRADLE uses for
correctBias_stored.

Adds information to bin/cradle and the README to explain the best way to name
the output directory.

Sets write index for hdf5 files