Improves computation parallelization
This changes the parallelization to allow for discrete work units smaller
than a single chromosome. This should allow for better throughput in many
workloads.

In order for multiple processes to work on regions in the same chromosome we have to output
temp files. These are just pickled numpy objects.

CRADLE must not break chromosomes up across processors when combining temp files into the final
hdf5 files, though.

The algorithm for dividing up work is the same as the one used by correctBias_stored.

The output has been compared to the covariate command without this change and the resulting
values are exactly the same, even though the hdf5 files are different.