Set BINSIZE = 1 for CalculateCovariates
correctBias_stored requires BINSIZE to be 1 and may not work correctly
otherwise. This means that CalculateCovariates shouldn't support other bin sizes.

I've removed the option to change the binsize and made it a constant in
the code that uses it.

Additionally, since we know the binsize (1) I was able to simplify some code since
x * 1 = x
x % 1 = 0 (if x is an integer)