Removes gquadmax argument from correctBias command
Instead of having the user input the gquadmax value, the value is
calculated from the gquad files themselves. Each file has a header
that contains the maxVal in the file, rounded down to the nearest
integer. The largets maxVal across all the gquad files will be used
as the GQUAD_MAX value.