Move constants around to memoize covariate funcs
lru_cache doesn't work if any of the arguments aren't hashable and
neither dictionaries nor lists are hashable. The big lists of MGW and
PROT values were moved into covariateUtils.pyx so they don't have to
get passed in via function args and the functions can continue to be
memoized with lru_cache.

The MGW and PROT lists were simplified to improve compilation times --
a plain python list that large was taking forever to compile.