Inline "findStartGibbs"
This huge unweildy nonsense replaces a call to cu.findStartGibbs.
findStartGibbs was a huge bottleneck to speed so this code inlines
and unravels it a bit. Additionally, sequence has been changed elsewhere
to an array of bytes (instead of unicode characters). This allows us to
treat it like an array of chars for fast tests and access.

On the benchmark I'm using it brings the time to calculate covariates
down from 12 seconds to 2 seconds. No other change comes close to being as important.