Fix extends, remove casts, avoid delete
There was a lot of extending a list by a list of length one. This
should be an append instead.

found some places with unnecessary casting -- the values were already
thre right type.

One of the algorithms iterated through a list, keeping track of
items to delete later (using numpy). More simply is to create a new
list and just not add the bad items.