Combine defineRegion and doWindowApproach into one
The output of defineRegion is used as the input of doWindowApproach.
However, these two functions were called using multiprocessing separately
so there was a step involved of writing the output out and reading it in that
was unnecessary. I removed this step. I also removed another unnecessary "write output
to a file" step and instead just pass a list of filenames around.