Fixes issue in callpeak setAnalysisRegion
In an older version of the code "pos" was an index into overlappedBL. A
change was made so overlappedBL was iterated over directly and pos
became a small numpy array. But it was still used as a number in one place!

So this change iterates over overlappedBL using enumerate so pos can be
the number it was supposed to be and blRegion is the value at that index.